Splits engine and error
